% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_network_errors.r
\name{check_network_errors}
\alias{check_network_errors}
\title{Check for network errors in the RSelenium session}
\usage{
check_network_errors(rmdr)
}
\arguments{
\item{rmdr}{An `RSelenium` Remote WebDriver object.}
}
\value{
A `logical` value:
\describe{
  \item{TRUE}{If a network error is found and the page has been refreshed.}
  \item{FALSE}{If no network error is detected.}
}
}
\description{
This internal helper function checks for known network errors in an RSelenium web session.
If an error is found, the page is refreshed.
}
\details{
The function looks for typical network error messages on the webpage and refreshes
the page if any of these errors occur. If no errors are found, the function ends without action.
}
