% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_back.r
\name{click_back}
\alias{click_back}
\title{Clicks the "Back" button of a webpage}
\usage{
click_back(rmdr, max_attempts = 10)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{max_attempts}{The maximum number of attempts to successfully click the button. The default value is 10.}
}
\value{
Returns `NULL` since the function does not provide any values. On success, a message is printed indicating that the Back button was clicked. In case of repeated failure, an error message is displayed.
}
\description{
This function attempts to click the "Back" button of a webpage to return to the previous page or overview.
It makes several attempts in case of errors and checks for network issues to ensure the button is clicked.
}
