% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_course_url.r
\name{click_course_url}
\alias{click_course_url}
\title{Click on a course URL and check page transition}
\usage{
click_course_url(i, rmdr, max_attempts = 10)
}
\arguments{
\item{i}{A `character` string with the CSS selector for the course URL.}

\item{rmdr}{A `remoteDriver` object from `RSelenium` that controls the browser instance.}

\item{max_attempts}{An `integer` specifying the maximum number of attempts (default: 10).}
}
\value{
Does not return a value but prints console messages to document the progress.
}
\description{
This function searches for a course url, scrolls it into view,
extracts the title, clicks on the element, and checks if the transition
to the course page was successful. If not, it checks whether the `remoteDriver`
is still on the overview page and attempts to navigate again if necessary.
}
