% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_next_page.r
\name{click_next_page}
\alias{click_next_page}
\title{Click on the "Next" button to display the next overview page}
\usage{
click_next_page(
  rmdr,
  css_next_page = "#genSearchRes\\\\:id3f3bd34c5d6b1c79\\\\:id3f3bd34c5d6b1c79Navi2next",
  max_attempts = 10
)
}
\arguments{
\item{rmdr}{An `RSelenium` Remote WebDriver object.}

\item{css_next_page}{A character string specifying the CSS selector of the "Next" button.}

\item{max_attempts}{The maximum number of attempts to find and click the "Next" button.
Default: `10`.}
}
\value{
No return value. The function interacts with the `RSelenium` WebDriver and controls
  navigation within the web application.
}
\description{
This function searches for the "Next" button on an overview page and clicks on it.
If the button is not found or a network issue occurs, several attempts are made to load the next page.
If the next page cannot be loaded successfully, it checks whether the WebDriver is still on an event page.
}
\details{
The function proceeds as follows:
- Checks for potential network issues and reloads the page if necessary.
- Scrolls down and looks for the "Next" button.
- If the button is found, it clicks it and reloads the page.
- Checks if the next overview page was successfully reached by searching for an element with
  search results.
- If the next page is not successfully loaded, it checks if the WebDriver is still on an event page.
  If so, it goes back to the previous page.
- If the WebDriver is neither on a course page nor an overview page, the process is aborted.
}
\seealso{
[check_network_errors()], [click_back()]
}
