% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown_sleep.r
\name{countdown_sleep}
\alias{countdown_sleep}
\title{Display a countdown with remaining seconds}
\usage{
countdown_sleep(seconds)
}
\arguments{
\item{seconds}{A numeric value specifying the number of seconds the countdown should last.}
}
\description{
This function performs a countdown and displays the remaining seconds in the console. The output is shown in a single line, which is updated every second to inform the user about the remaining time until the next step.
}
\details{
The countdown is displayed in the console, with each remaining second shown as an updated number. Once the countdown is finished, a message is displayed indicating that the waiting time is over.
}
