% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url_selectors.r
\name{generate_url_selectors}
\alias{generate_url_selectors}
\title{Generates CSS selectors for course links}
\usage{
generate_url_selectors(
  rmdr,
  selector_start = 0,
  css_first_selector,
  css_max_selector =
    "#genSearchRes\\\\:id3f3bd34c5d6b1c79\\\\:id3f3bd34c5d6b1c79Navi2_div > div > span.dataScrollerResultText"
)
}
\arguments{
\item{rmdr}{An RSelenium remote driver object to control the web browser.}

\item{selector_start}{A numeric value indicating the starting index of the selectors (default: 0).}

\item{css_first_selector}{A format string for the CSS selector of the course links, with \code{\%d} as a placeholder for the index.}

\item{css_max_selector}{A CSS selector to extract the maximum number of search results.}
}
\value{
A vector of CSS selectors for the course links.
}
\description{
This function creates a list of CSS selectors that can be used to select course links
on a webpage. The maximum number of selectors is extracted from the total number of search results.
}
