% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_to_last_active_page.r
\name{go_to_last_active_page}
\alias{go_to_last_active_page}
\title{Navigate to the target page based on a selector}
\usage{
go_to_last_active_page(rmdr, selector)
}
\arguments{
\item{rmdr}{An RSelenium driver object used for interaction with the webpage.}

\item{selector}{A numeric value representing the target selector.}
}
\value{
Does not return any values. The function navigates through the pages and prints messages
to document the progress.
}
\description{
This function calculates how many pages need to be flipped through to reach a specific page,
determined by the `selector` value. If the selector is greater than 100, navigation occurs in steps of 10,
otherwise in steps of 1.
}
\seealso{
\code{\link{click_next_page}}, \code{\link{click_next_10_pages}}
}
