% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_column_names_from_values.r
\name{remove_column_names_from_values}
\alias{remove_column_names_from_values}
\title{Removes column names from the cells of a DataFrame}
\usage{
remove_column_names_from_values(df)
}
\arguments{
\item{df}{A DataFrame from which the column names should be removed from the cells.}
}
\value{
A DataFrame with the column names removed from the cells.
}
\description{
This function removes the column names from the cell values of the respective column
in a DataFrame. It is used as an internal helper function.
}
\keyword{internal}
