% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_base_info.r
\name{scrape_base_info}
\alias{scrape_base_info}
\title{Scrape Base Information from Multiple Pages}
\usage{
scrape_base_info(rmdr, css_max_selector)
}
\arguments{
\item{rmdr}{Ein WebDriver-Objekt, das für die Interaktion mit der Webseite verwendet wird. Wird durch `RSelenium` erzeugt.}

\item{css_max_selector}{Ein CSS-Selektor, der auf das Element zeigt, das die maximale Seitenzahl enthält
die zum Scraping verwendet wird.}
}
\value{
Ein `tibble`, das die extrahierten Tabellen von allen Seiten enthält.
        Jede Tabelle wird zu einer `data.frame` innerhalb des `tibble` zusammengefügt.
}
\description{
Diese Funktion extrahiert Tabellen von mehreren Seiten einer Webseite, die mit einem CSS-Selektor definiert sind. 
Sie liest den HTML-Code der Tabelle, extrahiert ihn mit `rvest` und fügt ihn zu einem `tibble` zusammen. Der 
Prozess wiederholt sich für alle Seiten, die durch den maximalen Seitenwert bestimmt werden.
}
