% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_for_missing_data.r
\name{scrape_for_missing_data}
\alias{scrape_for_missing_data}
\title{Scrape fehlende Kursdaten aus dem HIS-System}
\usage{
scrape_for_missing_data(rmdr, missing_data, num_sem_selector)
}
\arguments{
\item{rmdr}{Ein `RSelenium::remoteDriver`-Objekt, das mit dem Browser verbunden ist.}

\item{missing_data}{Ein `tibble` mit den Spalten `titel` und `nummer`, die gescraped werden sollen.}

\item{num_sem_selector}{Eine Zeichenkette oder Zahl zur Auswahl des Semesters (z. B. `"2"`).}
}
\value{
Ein `tibble` mit den gesammelten Kursinformationen.
}
\description{
Führt ein halbautomatisiertes Scraping für Kurse durch, deren Informationen fehlen,
basierend auf Titel und Nummer. Die Ergebnisse werden in einer RDS-Datei gespeichert.
}
