% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_his.r
\name{scrape_his}
\alias{scrape_his}
\title{Scraping function for the HIS system}
\usage{
scrape_his(rmdr, chunks, uni_name)
}
\arguments{
\item{rmdr}{An `RSelenium` RemoteDriver object used for the scraping.}

\item{chunks}{A list of vectors containing the URLs or selectors to be scraped.}

\item{uni_name}{Name of the University. Important for the filename in the context from data export.}
}
\value{
A `tibble` containing the collected course data. If the function unexpectedly terminates,
the data collected up to that point will be returned.
}
\description{
This function extracts course information from the HIS system by visiting a given list of URLs
and extracting relevant content such as dates, content, modules, and degree programs.
}
\details{
The function iterates over the `chunks` and calls various scraping functions for each URL or selector:
- `scrape_termine()`: Extracts the course dates.
- `scrape_inhalte()`: Extracts the course content.
- `scrape_module()`: Extracts the associated modules.
- `scrape_studiengaenge()`: Extracts the associated degree programs.

If an error occurs, it is caught and logged. The function continues running to collect as much
data as possible.
}
