% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_inhalte.R
\name{scrape_inhalte}
\alias{scrape_inhalte}
\title{Scrape content from a tab}
\usage{
scrape_inhalte(
  rmdr,
  css_inhalte_tab =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:tabs\\\\:contentsTab > span:nth-child(1)",
  css_container =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:j_id_6s_13_2_\%d_1",
  css_title_selector =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:j_id_6s_13_2_\%d_1\\\\:collapsiblePanel > div.box_title > div > div.layoutFieldsetTitle.collapseTitle > h2",
  css_content_selector =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:j_id_6s_13_2_\%d_1\\\\:collapsiblePanel > div.box_content > fieldset",
  max_attempts = 10
)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{css_inhalte_tab}{A CSS selector for the "Content" tab. The default value is the selector for the "Content" tab.}

\item{css_container}{A CSS selector for the containers from which the data will be extracted. The default value is the selector for container elements.}

\item{css_title_selector}{A CSS selector for the titles of the containers. The default value is the selector for container titles.}

\item{css_content_selector}{A CSS selector for the content of the containers. The default value is the selector for container contents.}

\item{max_attempts}{The maximum number of attempts to successfully extract the content. The default value is 10.}
}
\value{
A tibble containing the extracted data from the "Content" tab. If no content is found, the function returns NULL.
}
\description{
This function searches a given RMD instance for the "Content" tab, clicks on it, and extracts data from the container elements found on the page. It makes several attempts in case of errors and returns the extracted data as a tibble.
}
