% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_module.r
\name{scrape_module}
\alias{scrape_module}
\title{Scrape modules/degree programs from a tab}
\usage{
scrape_module(
  rmdr,
  css_module_tab =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:tabs\\\\:modulesCourseOfStudiesTab > span:nth-child(1)",
  css_module_content =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:modules\\\\:moduleAssignments",
  max_attempts = 10
)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{css_module_tab}{The CSS selector for the tab containing the modules/degree programs.}

\item{css_module_content}{The CSS selector for the area displaying the modules/degree programs.}

\item{max_attempts}{The maximum number of attempts to successfully extract the data. The default value is 10.}
}
\value{
A tibble containing the extracted data from the "Modules/Degree Programs" tab. If no modules are found, the function returns `NULL`.
}
\description{
This function searches a given RMD instance for the "Modules/Degree Programs" tab, clicks on it, and extracts data from the corresponding HTML table. It makes several attempts in case of errors and returns the extracted data as a tibble. If no modules are found, it returns `NULL`.
}
