% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_studiengaenge.r
\name{scrape_studiengaenge}
\alias{scrape_studiengaenge}
\title{Scrape degree programs from the "Modules/Degree Programs" tab}
\usage{
scrape_studiengaenge(
  rmdr,
  css_zugeordnete_studiengaenge =
    "#detailViewData\\\\:tabContainer\\\\:term-planning-container\\\\:courseOfStudies\\\\:courseOfStudyAssignments\\\\:courseOfStudyAssignmentsTable",
  max_attempts = 10
)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{css_zugeordnete_studiengaenge}{The CSS selector for the HTML element containing the degree programs.}

\item{max_attempts}{The maximum number of attempts to successfully extract the degree programs. The default value is 10.}
}
\value{
A `tibble` containing the extracted degree program data. If no degree programs are found, an empty `tibble` is returned.
}
\description{
This function searches a given tab on a webpage for degree programs and extracts the relevant data into a `tibble`. It makes several attempts in case of errors and checks for network issues.
}
