% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_termine.r
\name{scrape_termine}
\alias{scrape_termine}
\title{Scrape dates from the "Dates" tab}
\usage{
scrape_termine(
  rmdr,
  css_labels = ".labelItemLine label",
  css_answers = ".labelItemLine .answer",
  max_attempts = 10
)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{css_labels}{A CSS selector for the labels to be extracted from the "Dates" tab. The default value is the selector for the label elements.}

\item{css_answers}{A CSS selector for the answers to be extracted from the "Dates" tab. The default value is the selector for the answer elements.}

\item{max_attempts}{The maximum number of attempts to successfully extract the data. The default value is 10.}
}
\value{
A tibble containing the extracted dates (labels and answers) from the "Dates" tab. If no data is found, the function returns an empty tibble.
}
\description{
This function searches a given RMD instance for the "Dates" tab, extracts the associated labels and answers, and returns them as a tibble. It makes several attempts in case of errors and checks for network issues.
}
