% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_semester_and_set_courses.R
\name{select_semester_and_set_courses}
\alias{select_semester_and_set_courses}
\title{Select semester and set courses}
\usage{
select_semester_and_set_courses(
  rmdr,
  base_url,
  num_sem_selector,
  num_courses,
  css_sem_dropdown,
  css_search_field,
  num_courses_selector
)
}
\arguments{
\item{rmdr}{An Rmdr object representing the Selenium WebDriver.}

\item{base_url}{The base URL to navigate to.}

\item{num_sem_selector}{The CSS selector for the semester to be selected.}

\item{num_courses}{The number of courses to be displayed on the page.}

\item{css_sem_dropdown}{The CSS selector for the dropdown menu to select the semester.}

\item{css_search_field}{The CSS selector for the "search terms" field.}

\item{num_courses_selector}{The CSS selector for the input field where the number of courses will be set.}
}
\value{
No return value. The function performs actions on the webpage.
}
\description{
This function navigates to a specified URL, selects a semester from a dropdown menu, and sets the number of courses displayed on the page. It adjusts the input field for the number of courses.
}
